Refactor Job code to allow better testability
